+++ Group Game Engine/Ship.h
@@ -9,7 +9,7 @@ class Ship :
 {
 public:
 	Ship();
-	Ship(int entityID, int lvl, std::string entityType, SDL_Texture* sprtsht, SDL_Rect spriteR, SDL_Rect hitBoxR, SDL_Renderer* ren);
+	Ship(float entityID, int lvl, std::string entityType, SDL_Texture* sprtsht, SDL_Rect spriteR, SDL_Rect hitBoxR, SDL_Renderer* ren);
 	~Ship();
 
 	virtual void control();
